requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
Babel==2.6.0
beautifulsoup4==4.7.1
bert==2.2.0
bert-tensorflow==1.0.1
blis==0.4.1
boto3==1.9.228
botocore==1.12.228
catalogue==1.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
cloudpickle==0.8.0
connexion==1.4.2
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==1.1.1
dill==0.2.8.2
distributed==1.25.3
dnspython==1.16.0
docutils==0.14
ds4biz-broker-manager==0.0.2.dev0
ds4biz-commons==0.0.2.dev0
ds4biz-flask==0.0.11
ds4biz-predictor-core==0.0.4.dev0
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
eli5==0.8.1
erlastic==2.0.0
eventlet==0.25.1
feedfinder2==0.0.4
feedparser==5.2.1
flashtext==2.7
Flask==1.0.2
Flask-Cors==3.0.6
flask-swagger-ui==3.18.0
fsspec==0.6.2
fuzzywuzzy==0.17.0
gast==0.2.2
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.20.1
h5py==2.9.0
HeapDict==1.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.5.0
inflection==0.3.1
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
kafka==1.3.5
kafka-python==1.4.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lightgbm==2.2.3
llvmlite==0.27.1
locket==0.2.0
lxml==4.3.1
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mock==3.0.5
monotonic==1.5
Morfessor==2.0.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
newspaper3k==0.2.8
nltk==3.4
numba==0.42.1
numpy==1.15.4
packaging==18.0
pandas==0.23.4
partd==0.3.9
pika==0.11.2
Pillow==5.4.1
plac==0.9.6
polyglot==16.7.4
preshed==3.0.2
protobuf==3.7.1
psutil==5.5.1
pycld2==0.31
Pygments==2.3.0
PyICU==2.3.1
pyparsing==2.3.1
Pyphen==0.9.5
pyrsistent==0.15.3
python-dateutil==2.7.5
python-Levenshtein==0.12.0
pytorch-pretrained-bert==0.6.2
pytz==2018.7
PyYAML==3.13
regex==2018.1.10
repoze.lru==0.7
requests==2.21.0
requests-file==1.4.3
s3fs==0.4.0
s3transfer==0.2.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.2.1
srsly==0.1.0
swagger-spec-validator==2.1.0
swifter==0.283
tabulate==0.8.3
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
textstat==0.5.6
thinc==7.1.1
tinysegmenter==0.3
tldextract==2.2.0
toolz==0.9.0
torch==1.2.0
tornado==5.1.1
tqdm==4.28.1
treetaggerwrapper==2.3
typing==3.6.6
ujson==1.35
urllib3==1.24.1
wasabi==0.2.2
Werkzeug==0.14.1
wrapt==1.10.11
xgboost==0.82
zict==0.1.3
zipp==2.1.0
